fitur tukarsampah all done
